+++ json_annotation/lib/src/json_serializable.dart
@@ -129,6 +129,15 @@ class JsonSerializable {
   /// ```
   final bool? createToJson;
 
+  /// Whether the generator should use UTC time for [DateTime] fields.
+  ///
+  /// If `true` (the default), all [DateTime] fields are written to JSON in
+  /// UTC.
+  ///
+  /// If a field is annotated with `JsonKey` with a non-`null` value for
+  /// `dateTimeUtc`, that value takes precedent.
+  final bool? dateTimeUtc;
+
   /// If `false` (the default), then the generated `FromJson` function will
   /// ignore unrecognized keys in the provided JSON [Map].
   ///
@@ -283,6 +292,7 @@ class JsonSerializable {
     this.converters,
     this.genericArgumentFactories,
     this.createPerFieldToJson,
+    this.dateTimeUtc,
   });
 
   factory JsonSerializable.fromJson(Map<String, dynamic> json) =>